feat: Add price history, prices, products, and promotions views with API integration
- Implemented PriceHistoryView to display price history data with pagination and filtering.
- Created PricesView to fetch and display price information, including country and store lookups.
- Developed ProductsView for managing product listings, including analytics and pricing details.
- Added PromotionsView for handling promotional data, including creation and deactivation of promotions.
- Introduced reference data functions for building product, country, and store lookups.
- Refactored API client to streamline HTTP requests and error handling across services.
- Enhanced error handling and logging for API interactions.
