+++ frontend/core/services/ai_chatbot_client.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
 import logging
-import time
 from typing import Any
 
-import requests
 from django.conf import settings
+
+from services.http import do_request
 
 logger = logging.getLogger("pricing_control_tower.frontend.ai_chatbot_client")
 
@@ -46,42 +46,28 @@ def ask_chatbot(
     """Send a question to the AI service /chat endpoint and return its JSON response."""
     url = f"{settings.AI_SERVICE_BASE_URL.rstrip('/')}/chat"
     payload = build_chat_payload(question, user_email, store_id)
-    start_time = time.perf_counter()
 
-    try:
-        response = requests.post(url, json=payload, timeout=30)
-        response.raise_for_status()
-        data = response.json()
-
-    except requests.exceptions.ConnectionError as exc:
-        log_ai_call_failure(user_email, "Unable to connect to AI service.")
-        raise AiChatbotConnectionError("Unable to connect to AI service.") from exc
-
-    except requests.exceptions.Timeout as exc:
-        log_ai_call_failure(user_email, "AI service request timed out.")
-        raise AiChatbotConnectionError("AI service request timed out.") from exc
-
-    except requests.exceptions.HTTPError as exc:
-        log_ai_call_failure(user_email, f"AI service returned an error: {response.status_code}")
-        raise AiChatbotResponseError(f"AI service returned an error: {response.status_code}") from exc
-
-    except requests.exceptions.JSONDecodeError as exc:
-        log_ai_call_failure(user_email, "AI service returned invalid JSON.")
-        raise AiChatbotResponseError("AI service returned invalid JSON.") from exc
-
-    logger.info(
-        "AI service call succeeded",
-        extra={
-            "extra_fields": {
-                "event": "ai_service_call_succeeded",
-                "status_code": response.status_code,
-                "duration_ms": round((time.perf_counter() - start_time) * 1000, 2),
-                "user_email": user_email,
-            }
-        },
+    return do_request(
+        "POST",
+        url,
+        service_label="AI service",
+        connection_error_cls=AiChatbotConnectionError,
+        response_error_cls=AiChatbotResponseError,
+        json=payload,
+        timeout=30,
+        on_success=lambda status_code, duration_ms: logger.info(
+            "AI service call succeeded",
+            extra={
+                "extra_fields": {
+                    "event": "ai_service_call_succeeded",
+                    "status_code": status_code,
+                    "duration_ms": duration_ms,
+                    "user_email": user_email,
+                }
+            },
+        ),
+        on_failure=lambda status_code, duration_ms, error: log_ai_call_failure(user_email, error),
     )
-
-    return data
 
 
 def log_ai_call_failure(user_email: str | None, error: str) -> None:
