+++ frontend/core/tests.py
@@ -1,3 +1,4 @@
+import time
 from unittest.mock import patch
 
 from django.contrib.auth.models import User
@@ -24,7 +25,7 @@ class ChatbotViewTests(TestCase):
             HTTP_X_REQUESTED_WITH="XMLHttpRequest",
         )
 
-    @patch("core.views.ask_chatbot")
+    @patch("core.views.chatbot.ask_chatbot")
     def test_successful_answer_is_stored_and_returned(self, mock_ask_chatbot):
         mock_ask_chatbot.return_value = {
             "answer": "Un store manager peut consulter son magasin.",
@@ -45,7 +46,7 @@ class ChatbotViewTests(TestCase):
         self.assertEqual(history[-2], {"role": "user", "content": "Que peut faire un store manager ?"})
         self.assertEqual(history[-1]["status"], "routed")
 
-    @patch("core.views.ask_chatbot")
+    @patch("core.views.chatbot.ask_chatbot")
     def test_connection_error_returns_clean_message_no_internals(self, mock_ask_chatbot):
         mock_ask_chatbot.side_effect = AiChatbotConnectionError("Connection refused")
 
@@ -57,7 +58,7 @@ class ChatbotViewTests(TestCase):
         self.assertIn("momentanément indisponible", data["content"])
         self.assertNotIn("Connection refused", data["content"])
 
-    @patch("core.views.ask_chatbot")
+    @patch("core.views.chatbot.ask_chatbot")
     def test_response_error_returns_clean_message_no_internals(self, mock_ask_chatbot):
         mock_ask_chatbot.side_effect = AiChatbotResponseError("JSONDecodeError: boom")
 
@@ -68,7 +69,7 @@ class ChatbotViewTests(TestCase):
         self.assertIn("réponse inattendue", data["content"])
         self.assertNotIn("JSONDecodeError", data["content"])
 
-    @patch("core.views.ask_chatbot")
+    @patch("core.views.chatbot.ask_chatbot")
     def test_unexpected_exception_returns_generic_message_no_traceback(self, mock_ask_chatbot):
         mock_ask_chatbot.side_effect = ValueError("unexpected internal bug")
 
@@ -80,12 +81,120 @@ class ChatbotViewTests(TestCase):
         self.assertNotIn("unexpected internal bug", data["content"])
         self.assertNotIn("Traceback", data["content"])
 
-    @patch("core.views.ask_chatbot")
+    @patch("core.views.chatbot.ask_chatbot")
     def test_empty_question_does_not_call_ai_service(self, mock_ask_chatbot):
         response = self.post_ajax("   ")
 
         self.assertEqual(response.status_code, 400)
         mock_ask_chatbot.assert_not_called()
+
+
+class ChatbotHistoryLimitsTests(TestCase):
+    """chatbot_history: per-message length cap and TTL-based expiry (may contain pricing
+    data surfaced by tool answers, so it shouldn't linger indefinitely)."""
+
+    def setUp(self):
+        self.user = User.objects.create_user(
+            username="history_limits_user",
+            password="Password123!",
+            email="history.limits@pct.local",
+        )
+        self.client.force_login(self.user)
+        self.url = reverse("core:chatbot")
+
+    def post_ajax(self, message):
+        return self.client.post(
+            self.url,
+            {"message": message},
+            HTTP_X_REQUESTED_WITH="XMLHttpRequest",
+        )
+
+    @patch("core.views.chatbot.ask_chatbot")
+    def test_long_user_message_is_truncated_before_storage(self, mock_ask_chatbot):
+        mock_ask_chatbot.return_value = {
+            "answer": "ok",
+            "status": "routed",
+            "selected_tool": None,
+            "metadata": {},
+        }
+
+        long_message = "a" * 5000
+        self.post_ajax(long_message)
+
+        history = self.client.session["chatbot_history"]
+        stored_question = history[-2]["content"]
+        self.assertEqual(len(stored_question), 4001)  # MAX_MESSAGE_LENGTH + ellipsis
+        self.assertTrue(stored_question.endswith("…"))
+
+    @patch("core.views.chatbot.ask_chatbot")
+    def test_long_assistant_reply_is_truncated_before_storage(self, mock_ask_chatbot):
+        mock_ask_chatbot.return_value = {
+            "answer": "b" * 5000,
+            "status": "routed",
+            "selected_tool": None,
+            "metadata": {},
+        }
+
+        self.post_ajax("question")
+
+        history = self.client.session["chatbot_history"]
+        stored_answer = history[-1]["content"]
+        self.assertEqual(len(stored_answer), 4001)
+        self.assertTrue(stored_answer.endswith("…"))
+
+    @patch("core.views.chatbot.ask_chatbot")
+    def test_post_refreshes_history_updated_at(self, mock_ask_chatbot):
+        mock_ask_chatbot.return_value = {
+            "answer": "ok",
+            "status": "routed",
+            "selected_tool": None,
+            "metadata": {},
+        }
+
+        self.post_ajax("question")
+
+        self.assertIn("chatbot_history_updated_at", self.client.session)
+
+    @patch("core.views.chatbot.ask_chatbot")
+    def test_expired_history_is_cleared_on_next_post(self, mock_ask_chatbot):
+        session = self.client.session
+        session["chatbot_history"] = [{"role": "user", "content": "old pricing question"}]
+        session["chatbot_history_updated_at"] = time.time() - 3601  # just past the 1h TTL
+        session.save()
+
+        mock_ask_chatbot.return_value = {
+            "answer": "fresh answer",
+            "status": "routed",
+            "selected_tool": None,
+            "metadata": {},
+        }
+
+        self.post_ajax("new question")
+
+        history = self.client.session["chatbot_history"]
+        # only the new turn should remain — the expired one was dropped, not carried forward
+        self.assertEqual(len(history), 2)
+        self.assertEqual(history[0]["content"], "new question")
+
+    def test_expired_history_is_cleared_on_page_load(self):
+        session = self.client.session
+        session["chatbot_history"] = [{"role": "user", "content": "old pricing question"}]
+        session["chatbot_history_updated_at"] = time.time() - 3601
+        session.save()
+
+        response = self.client.get(self.url)
+
+        self.assertEqual(response.context["chatbot_history"], [])
+
+    def test_fresh_history_is_kept_on_page_load(self):
+        session = self.client.session
+        session["chatbot_history"] = [{"role": "user", "content": "recent question"}]
+        session["chatbot_history_updated_at"] = time.time() - 60
+        session.save()
+
+        response = self.client.get(self.url)
+
+        self.assertEqual(len(response.context["chatbot_history"]), 1)
 
 
 # ---------------------------------------------------------------------------
