+++ frontend/core/views/home.py
@@ -1,0 +1,6 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views.generic import TemplateView
+
+
+class HomeView(LoginRequiredMixin, TemplateView):
+    template_name = "core/home.html"
